End of Module

+++ controllers/products.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const {validationResult } = require('express-validator');
 
 exports.getProducts = (req, res, next) => {
     Product.find({}).lean()
@@ -35,12 +36,16 @@ exports.getProduct = (req, res, next) => {
 
 exports.getProductEdit = (req, res, next) => {
     const productId = req.params.productId;
-    Product.findById(productId).lean()
+    const userId = req.session.user._id;
+    Product.findOne({_id: productId, userId: userId}).lean()
         .then(product => {
             res.render('templates/edit-product', {
                 product: product,
                 pageTitle: product.title,
-                path: `/admin/edit-product`
+                path: `/admin/edit-product`,
+                errorMessage: null,
+                validationErrors: null,
+                oldInputs: null
             });
         })
         .catch(e => {
@@ -51,20 +56,43 @@ exports.getProductEdit = (req, res, next) => {
 
 exports.postProductEdit = (req, res, next) => {
     const {productId, ...updatedProduct} = req.body;
-    console.log(updatedProduct);
-    Product.updateOne({_id: productId}, updatedProduct)
-        .then(() => {
-            res.redirect("/admin/products");
-        })
-        .catch(e => {
-            console.log(e);
-            res.redirect('/');
+    const userId = req.session.user._id;
+    const result = validationResult(req);
+    if (!result.isEmpty()) {
+        res.status(422).render('templates/edit-product', {
+            path: `/admin/edit-product`,
+            pageTitle: 'Edit Product',
+            errorMessage: result.array()[0].msg,
+            oldInputs: {
+                title: req.body.title,
+                imageUrl: req.body.imageUrl,
+                price: req.body.price,
+                description: req.body.description,
+                productId: productId
+            },
+            validationErrors: result.array()
         });
+    }
+    else {
+        Product.updateOne({_id: productId, userId: userId}, updatedProduct)
+            .then(response => {
+                if (response.matchedCount === 0) {
+                    console.log("Cannot edit product");
+                    return res.redirect('/');
+                }
+                res.redirect("/admin/products");
+            })
+            .catch(e => {
+                console.log(e);
+                res.redirect('/');
+            });
+    }
 };
 
 exports.postProductDelete = (req, res, next) => {
-    Product.deleteOne({_id: req.body.productId})
-        .then(() => {
+    const userId = req.session.user._id;
+    Product.deleteOne({_id: req.body.productId, userId: userId})
+        .then(result => {
             res.redirect("/admin/products");
         })
         .catch(e => {
@@ -74,7 +102,8 @@ exports.postProductDelete = (req, res, next) => {
 };
 
 exports.getAdminProducts = (req, res, next) => {
-    Product.find({}).lean()
+    const userId = req.session.user._id;
+    Product.find({userId: userId}).lean()
         .then(products => {
             res.render('templates/products', {
                 products: products,
@@ -85,20 +114,40 @@ exports.getAdminProducts = (req, res, next) => {
 };
 
 exports.getAddProduct = (req, res, next) => {
-    res.render('templates/add-product', {
+    res.status(200).render('templates/add-product', {
+        path: `/admin/add-product`,
         pageTitle: 'Add Product',
-        path: '/admin/add-product'
+        errorMessage: null,
+        oldInputs: null,
+        validationErrors: null
     });
 };
 
 exports.postAddProduct = (req, res, next) => {
-    const product = new Product(req.body);
-    product.userId = req.session.user._id;
-    product.save()
-        .then(result => {
-            res.status(201).redirect('/');
-        })
-        .catch(err => {
-            console.log(err);
-        })
+    const result = validationResult(req);
+    if (!result.isEmpty()){
+        res.status(422).render('templates/add-product', {
+            path: `/admin/add-product`,
+            pageTitle: 'Add Product',
+            errorMessage: result.array()[0].msg,
+            oldInputs: {
+                title: req.body.title,
+                imageUrl: req.body.imageUrl,
+                price: req.body.price,
+                description: req.body.description,
+            },
+            validationErrors: result.array()
+        });
+    } else {
+        const product = new Product(req.body);
+        product.userId = req.session.user._id;
+        product.save()
+            .then(result => {
+                res.status(201).redirect('/');
+            })
+            .catch(err => {
+                console.log(err);
+            })
+    }
+
 };
